DISMTools 0.6.1 Update 1 (#235) (#236)
* [Installer] Update Inno Setup to 6.4.1

* Update assembly information

* [Fix] Control window resize events

* [Fix] Skip second scan of main package from appinstaller

* [PE Helper] Improvements, lots of improvements

- Add SecureStartup and EnhancedStorage OCs to PE images, mainly to deal with BitLocker-encrypted drives
- Improved output of package addition and OC copy

* [PE Helper] Add Arm64EC compatibility

* [PE Helper] Add experimental ARM64 builds of DIM

* [PE Helper] Official ARM64 support for DIM

The ARM64 build of the Driver Installation Module (DIM) works without any problems on a Raspberry Pi. This officially introduces support for that architecture on the PE Helper when it creates the ISO files.

The Arm64EC version is removed because of 2 reasons:

1. It didn't work
2. It isn't necessary, since we don't have to slowly convert the AMD64 native code to ARM64 and have an in-between version

However, ARM64 Windows PEs will still have support for Arm64EC, for **your** applications (you will have to target them to this platform first)

An additional note is that, in order to fully experience the DIM in the ARM64 DISMTools PE, you will have to start ISO creation from within an ARM64 system (so use your Copilot+ PCs with Recall and other stuff no one asked, or Project Volterra, or your Lumia phone, or your Android phone... WoA is everywhere! When will it arrive to iPhones, though?)

* [Enhance] Added support for installations in removable media

* [PE Helper] Add status label to DIM

* [Fix] Fixed product key validation algo

* [PE Helper] ARM64 restart dialog

* [Updater] Change location of update files

* [Unattended answer file] Update UnattendGen, disable password on open network connections

* [Unattended answer file] Fixed access violation exception

* [REL] Update What's New section, and Update Info files