[PE Helper/PXE Helpers] Add something interesting
This is a teaser for a new provider compatibility helper for PXE. More info when I start development on 0.7.1.

The files themselves are already done, but the following needs to be done:

- More testing

That's why, instead of seeing the full thing, you'll see the teaser. You'll get to experience and provide feedback on this during the development of 0.7.1