[ServiceManager] Add basic device/service lister
The system device/service manager interface is in EXTREMELY EARLY stages and will only list the results obtained. No management actions have been implemented yet.

There are also several bugs in the manager that need to be fixed, and the overall manager UI also needs to adapt to the color schemes.

Regardless, this is the INTRODUCTION of a new thing.