[ServiceManager] System Service Manager (2025/09/09)
A property section has been added to the manager. Parsing of indirect strings has also been implemented.

None of the properties can be changed yet (and it will most likely be like that for 0.7.1 P1). Other quirks need to be addressed as well. Anyways, this is taking shape.

Also, happy 30th to one of the franchises that has become quite the talk recently: https://en.wikipedia.org/wiki/Rayman