DISMTools 0.7.1 Preview 1 (#281)
* [PE Helper/AutoRun] New Autorun application

* [PE Helper/PXE Helpers] Add network adapter install screen, add FOG Helper

* [PE Helper] Add Sysprep Preparator, modify autorun

* [PE Helper/SysprepPreparator] Update binary

Check out https://github.com/CodingWonders/SysprepPreparator for more information about commits.

* [Unattended answer file/ADDS] Add OU/User Picker

* [Unattended answer file/ADDS] Add DynaLog to new functions, add enabled checks

* [Unattended answer file/ADDS] Replace account disable value with constant

* [Unattended answer file] Update UnattendGen

* [Unattended answer file] Native support for Batch, more goodies

Thanks to @cschneegans for suggesting this. Yes, a lot had to be made, but for the sake of getting something beyond good.
- Batch scripts are now imported NATIVELY thanks to the updated UnattendGen adding native support for batch scripts, so we no longer need the wrapper
- Multiple scripts can now be specified for each stage. Rather than supporting only 3 or 4 scripts, it allows adding as many as you prefer (maybe you need 10+?)
- As a nice QoL feature, syntax highlighting is now dynamically adjusted for the selected scripting language
*NOTE:* VBScript and JScript are not supported yet. They might in the future. Registry is not planned to be added yet as well because there are tools in DT that allow you to customize the target system's registry, though the enum value is available.

* [PE Helper] Make some settings permanent

* [PE Helper/SysprepPreparator] Update binary

Check out https://github.com/CodingWonders/SysprepPreparator for more information about commits.

* [PE Helper/SysprepPreparator] Update binary

----- Initial Release of Sysprep Preparation Tool (v0.7.1.2591) -----

Check out https://github.com/CodingWonders/SysprepPreparator for more information about commits.

* [ServiceManager] Add basic device/service lister

The system device/service manager interface is in EXTREMELY EARLY stages and will only list the results obtained. No management actions have been implemented yet.

There are also several bugs in the manager that need to be fixed, and the overall manager UI also needs to adapt to the color schemes.

Regardless, this is the INTRODUCTION of a new thing.

* [ServiceManager] System Service Manager (2025/09/09)

A property section has been added to the manager. Parsing of indirect strings has also been implemented.

None of the properties can be changed yet (and it will most likely be like that for 0.7.1 P1). Other quirks need to be addressed as well. Anyways, this is taking shape.

Also, happy 30th to one of the franchises that has become quite the talk recently: https://en.wikipedia.org/wiki/Rayman

* [Library] Update Markdig to 0.42.0

* [ServiceManager] System Service Manager (2025/09/11)

The System Service Manager has seen several improvements, such as an icon, dark mode support, and so much more.

This will be the implementation of the System Service Manager for Preview 1.

* [REL] Update What's New Section